Move logic part from today.js to logic.js
